requirements.txt
torch>=2.2.0
transformers==4.34.0
faiss-cpu==1.7.4
PyMuPDF==1.23.8
numpy>=1.24.0
tqdm==4.66.1
pillow==10.2.0
streamlit==1.31.1
accelerate==0.27.2
diffusers==0.19.3
huggingface-hub==0.16.4
opencv-python-headless==4.8.0.76